removed useless files
